[nan-1762] remove microsoft cast
